pushed readme.md to help clarify meaning
In /, the lack of information created confusion. Fortunately, this update resolves the glaring formatting inconsistency by taking the open-source liberty of adding a clarifying readme file to help clear up confusion and help newcomers understand our mission here at reductio. The commit(s) in the PR affect only the +readme.md file, increasing readability.

Indeed, this update is vital to the repository because just a small investment in the maintainability of a codebase makes it easier to pick it up after a break and can provide you with an insurance policy should the source from this relatively disposable software turn out more useful than originally thought.